using int rather than uin64_t for emitted signal - used to work but seem to be qt version dependent

+++ source/daqlite/MainWindow.h
@@ -44,7 +44,7 @@ public slots:
   void handleInvertButton();
   void handleAutoScaleXButton();
   void handleAutoScaleYButton();
-  void handleKafkaData(uint64_t ElapsedCountNS);
+  void handleKafkaData(int ElapsedCountNS);
 
 private:
   Ui::MainWindow *ui;
